+++ R/sample_grid.R
@@ -56,6 +56,7 @@ sample_grid_scan <- function(scan_results,
   
   # recreate parameters for re running
   param_grid <- data.frame("beta" = beta, "start_date" = dates)
+  model <- scan_results$inputs$model
   model_params <- scan_results$inputs$model_params
   pars_obs <- scan_results$inputs$pars_obs
   data <- scan_results$inputs$data
@@ -68,11 +69,10 @@ sample_grid_scan <- function(scan_results,
   traces <- furrr::future_pmap(
     .l = param_grid, .f = beta_date_particle_filter,
     generate_beta_func = scan_results$generate_beta_func,
-    model_params = model_params, data = data, 
+    model = model, model_params = model_params, data = data,
     pars_obs = pars_obs, n_particles = n_particles,
     forecast_days = forecast_days, 
-    save_particles = TRUE, return = "sample"
-  )
+    save_particles = TRUE, return = "sample")
 
   # collapse into an array of trajectories
   # the trajectories are different lengths in terms of dates
@@ -100,7 +100,7 @@ sample_grid_scan <- function(scan_results,
                 model_params = model_params,
                 pars_obs = pars_obs,
                 data = data,
-                model = sircovid(params = model_params)))
+                model = model))
   
   class(res) <- "sample_grid_search"