+++ tests/testthat/test-grid_search.R
@@ -26,10 +26,11 @@ test_that("Small grid search works", {
    
   expect_is(scan_results, "sircovid_scan")
   expect_true("inputs" %in% names(scan_results))
-  expect_true(all(names(scan_results$inputs) == c("model_params","pars_obs","data")))
+  expect_setequal(names(scan_results$inputs),
+                  c("model", "model_params", "pars_obs", "data"))
   
-  beta_grid = seq(min_beta, max_beta, beta_step)
-  date_grid = seq(as.Date(first_start_date), as.Date(last_start_date), day_step)
+  beta_grid <- seq(min_beta, max_beta, beta_step)
+  date_grid <- seq(as.Date(first_start_date), as.Date(last_start_date), day_step)
   expect_equal(scan_results$x, beta_grid)
   expect_equal(scan_results$y, date_grid)
   expect_equal(dim(scan_results$renorm_mat_LL), dim(scan_results$mat_log_ll))
@@ -39,6 +40,49 @@ test_that("Small grid search works", {
   # Plots run, but not checked
   plot(scan_results)
 })
+
+test_that("Small grid search works with new model", {
+  set.seed(1)
+
+  data <- read.csv(sircovid_file("extdata/example.csv"),
+                   stringsAsFactors = FALSE)
+ 
+  # Parameters for run
+  min_beta <- 0.1
+  max_beta <- 0.2
+  beta_step <- 0.05
+  first_start_date <- "2020-01-21"
+  last_start_date <- "2020-01-22"
+  day_step <- 1
+  
+  skip("TODO: Add patched parameters here")
+  scan_results = scan_beta_date(
+    min_beta = min_beta,
+    max_beta = max_beta,
+    beta_step = beta_step,
+    first_start_date = first_start_date, 
+    last_start_date = last_start_date, 
+    day_step = day_step,
+    data = data,
+    model = "new_hospital_model")
+
+  expect_is(scan_results, "sircovid_scan")
+  expect_true("inputs" %in% names(scan_results))
+  expect_setequal(names(scan_results$inputs),
+                  c("model", "model_params", "pars_obs", "data"))
+  
+  beta_grid <- seq(min_beta, max_beta, beta_step)
+  date_grid <- seq(as.Date(first_start_date), as.Date(last_start_date), day_step)
+  expect_equal(scan_results$x, beta_grid)
+  expect_equal(scan_results$y, date_grid)
+  expect_equal(dim(scan_results$renorm_mat_LL), dim(scan_results$mat_log_ll))
+  expect_equal(dim(scan_results$renorm_mat_LL), c(length(beta_grid), length(date_grid)))
+  expect_true(all(scan_results$renorm_mat_LL < 1 & scan_results$renorm_mat_LL > 0))
+  
+  # Plots run, but not checked
+  plot(scan_results)
+})
+
 
 test_that("Transmission is more likely", {
   set.seed(1)