+++ tests/testthat/test-sample_grid.R
@@ -46,9 +46,10 @@ test_that("sample_grid_scan works", {
                           n_sample_pairs = n_sample_pairs, 
                           n_particles = 10)
 
+  model <- res$inputs$model$model(user = res$inputs$model_params)
   # check length based on model and dates
   days_between <- length( min(as.Date(res$param_grid$start_date)) : as.Date(tail(rownames(res$trajectories[,,1]),1)))
-  expect_equal(dim(res$trajectories), c(days_between, length(res$inputs$model$initial()), n_sample_pairs))
+  expect_equal(dim(res$trajectories), c(days_between, length(model$initial()), n_sample_pairs))
   
   ## Testing plotting is always a nightmare
   if (TRUE) {